requirements.txt
email-validator>=2.3.0
flask-dance>=7.1.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.43
werkzeug>=3.1.3
flask-login>=0.6.3
oauthlib>=3.3.1
pyjwt>=2.10.1
requests>=2.32.0
fastapi>=0.114.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
flask-migrate>=4.0.0
